Added support for multiline ansible_managed (#32)
In the project in which this is being implemented, the ansible_managed
phrase spans over multiple lines. It explains that changes will not be
persisted, and where to find the ansible repository.

However, in the case of the templates generated here, the comment is
only applied to the first line of ansible managed. Accordingly, it
introduces various syntax errors, particularly in `/etc/profile.d/`.
This commmit uses the `comment` filter, introduced with Ansible, to
prefix all lines with a `#` character (or, the 'plain' filter).

Tested in the environment in which it previously failed.